updated transformers, tests and clean up

+++ worker/internal/benthos/transformers/generate_card_number.go
@@ -56,7 +56,7 @@ func GenerateCardNumber(luhn bool) (int64, error) {
 			return 0, err
 		}
 
-		returnValue = int64(val)
+		returnValue = val
 
 	}
 
@@ -83,7 +83,7 @@ func GenerateValidLuhnCheckCardNumber() (int64, error) {
 	if err != nil {
 		return 0, err
 	}
-	strNine := strconv.FormatInt(int64(nineDigits), 10)
+	strNine := strconv.FormatInt(nineDigits, 10)
 	for _, k := range strNine {
 		digit, err := strconv.Atoi(string(k))
 		if err != nil {
